+++ dip-template/runtimes/dip-provider/Cargo.toml
@@ -20,6 +20,7 @@ scale-info = {workspace = true, features = ["derive"]}
 # DIP
 did.workspace = true
 dip-support.workspace = true
+kilt-dip-support.workspace = true
 kilt-runtime-api-dip-provider.workspace = true
 pallet-did-lookup.workspace = true
 pallet-dip-provider.workspace = true
@@ -78,6 +79,7 @@ std = [
 	"scale-info/std",
   "did/std",
   "dip-support/std",
+  "kilt-dip-support/std",
   "kilt-runtime-api-dip-provider/std",
   "pallet-did-lookup/std",
   "pallet-dip-provider/std",
@@ -123,6 +125,7 @@ std = [
 ]
 runtime-benchmarks = [
   "did/runtime-benchmarks",
+  "kilt-dip-support/runtime-benchmarks",
   "pallet-did-lookup/runtime-benchmarks",
   "pallet-dip-provider/runtime-benchmarks",
   "pallet-web3-names/runtime-benchmarks",
