+++ pallets/pallet-dip-provider/src/traits.rs
@@ -53,9 +53,7 @@ pub use identity_dispatch::*;
 pub mod identity_dispatch {
 	use super::*;
 
-	use frame_support::{pallet_prelude::Get, weights::Weight};
-	use parity_scale_codec::Encode;
-	use sp_std::marker::PhantomData;
+	use frame_support::weights::Weight;
 
 	pub trait IdentityProofDispatcher<Identifier, IdentityRoot, AccountId, Details = ()> {
 		type PreDispatchOutput;
@@ -93,83 +91,6 @@ pub mod identity_dispatch {
 
 		fn dispatch(_pre_output: Self::PreDispatchOutput) -> Result<(), Self::Error> {
 			Ok(())
-		}
-	}
-
-	// Dispatcher using a type implementing the `SendXcm` trait.
-	// It properly encodes the `Transact` operation, then delegates everything else
-	// to the sender, similarly to what the XCM pallet's `send` extrinsic does.
-	pub struct XcmRouterDispatcher<Router, UniversalLocationProvider>(PhantomData<(Router, UniversalLocationProvider)>);
-
-	impl<Router, UniversalLocationProvider, Identifier, ProofOutput, AccountId, Details>
-		IdentityProofDispatcher<Identifier, ProofOutput, AccountId, Details>
-		for XcmRouterDispatcher<Router, UniversalLocationProvider>
-	where
-		Router: SendXcm,
-		UniversalLocationProvider: Get<InteriorMultiLocation>,
-		Identifier: Encode,
-		ProofOutput: Encode,
-		AccountId: Into<[u8; 32]> + Clone,
-	{
-		type PreDispatchOutput = Router::Ticket;
-		type Error = SendError;
-
-		fn pre_dispatch<Builder: TxBuilder<Identifier, ProofOutput, Details>>(
-			action: IdentityDetailsAction<Identifier, ProofOutput, Details>,
-			source: AccountId,
-			asset: MultiAsset,
-			weight: Weight,
-			destination: MultiLocation,
-		) -> Result<(Self::PreDispatchOutput, MultiAssets), Self::Error> {
-			// TODO: Replace with proper error handling
-			let dest_tx = Builder::build(destination, action)
-				.map_err(|_| ())
-				.expect("Failed to build call");
-
-			// TODO: Set an error handler and an appendix to refund any leftover funds to
-			// the provider parachain sovereign account.
-			let operation = [[
-				DescendOrigin(X1(AccountId32 {
-					network: None,
-					id: source.clone().into(),
-				})),
-				WithdrawAsset(asset.clone().into()),
-				BuyExecution {
-					fees: asset,
-					weight_limit: Limited(weight),
-				},
-				Transact {
-					origin_kind: OriginKind::Native,
-					require_weight_at_most: weight,
-					call: dest_tx,
-				},
-				RefundSurplus,
-				DepositAsset {
-					assets: Wild(All),
-					beneficiary: MultiLocation {
-						parents: 1,
-						// Re-anchor the same account junction as seen from the destination.
-						// TODO: Error handling
-						interior: Here
-							.into_location()
-							.reanchored(&destination, UniversalLocationProvider::get())
-							.unwrap()
-							.pushed_with_interior(AccountId32 {
-								network: None,
-								id: source.into(),
-							})
-							.unwrap()
-							.interior,
-					},
-				},
-			]]
-			.concat();
-			let op = Xcm(operation);
-			Router::validate(&mut Some(destination), &mut Some(op))
-		}
-
-		fn dispatch(pre_output: Self::PreDispatchOutput) -> Result<(), Self::Error> {
-			Router::deliver(pre_output).map(|_| ())
 		}
 	}
 }
