+++ runtimes/spiritnet/src/xcm_config.rs
@@ -27,7 +27,7 @@ use frame_support::{
 };
 use pallet_xcm::XcmPassthrough;
 use sp_core::ConstU32;
-use xcm::latest::prelude::*;
+use xcm::v3::prelude::*;
 use xcm_builder::{
 	AllowTopLevelPaidExecutionFrom, EnsureXcmOrigin, FixedWeightBounds, RelayChainAsNative, SiblingParachainAsNative,
 	SignedAccountId32AsNative, SignedToAccountId32, UsingComponents, WithComputedOrigin,
